+++ app/src/main/java/com/bucharest/qurio/presentation/home/HomeView.kt
@@ -1,0 +1,7 @@
+package com.bucharest.qurio.presentation.home
+
+import com.bucharest.qurio.presentation.base.BaseView
+
+interface HomeView : BaseView {
+    fun navigateToQuiz()
+}
