--- app/src/main/java/com/bucharest/qurio/ui/home/HomeView.kt
@@ -1,7 +1,0 @@
-package com.bucharest.qurio.ui.home
-
-import com.bucharest.qurio.base.BaseView
-
-interface HomeView : BaseView {
-    fun navigateToQuiz()
-}
