+++ app/src/main/res/navigation/nav_graph.xml
@@ -6,7 +6,7 @@
 
     <fragment
         android:id="@+id/homeFragment"
-        android:name="com.bucharest.qurio.ui.home.HomeFragment"
+        android:name="com.bucharest.qurio.presentation.home.HomeFragment"
         android:label="Home">
         <action
             android:id="@+id/action_homeFragment_to_quizFragment"
@@ -15,7 +15,7 @@
 
     <fragment
         android:id="@+id/quizFragment"
-        android:name="com.bucharest.qurio.ui.quiz.QuizFragment"
+        android:name="com.bucharest.qurio.presentation.quiz.QuizFragment"
         android:label="Quiz" />
 
 </navigation>
